58-Error-saving-viewable-of-type-torch_tensor-RuntimeError-result-type-Float-cant-be-cast-to-the-desired-output-type-Byte (#59)
* Convert torch tensors to float before saving

* bump version

+++ src/python/torch_tensor.py
@@ -31,7 +31,7 @@ try:
 
             pad_value = 255
             torchvision.utils.save_image(
-                obj,
+                obj.float(),
                 path,
                 normalize=normalize,
                 pad_value=pad_value,
